Yet another proto recrop  - Add State to ServicesProviders  - Move States to separate proto  - Remove instances reference from States  - Rename Statuses to States  - Fix paths
