+++ pkg/instances/proto/instances.proto
@@ -1,8 +1,24 @@
+/*
+Copyright © 2021-2022 Nikita Ivanovski info@slnt-opp.xyz
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
 syntax = "proto3";
 package nocloud.instances;
 
 import "google/protobuf/struct.proto";
 import "pkg/hasher/hasherpb/hasher.proto";
+import "pkg/states/proto/states.proto";
 
 message Instance {
     string uuid = 1;
@@ -12,7 +28,7 @@ message Instance {
     map<string, google.protobuf.Value> resources = 4
         [(nocloud.hasher.hashed) = true];
     map<string, google.protobuf.Value> data = 5;
-    State state = 6;
+    nocloud.states.State state = 6;
     string hash = 7 [(nocloud.hasher.hash) = true];
 }
 
@@ -41,44 +57,4 @@ message TestInstancesGroupConfigError {
 message TestInstancesGroupConfigResponse {
     bool result = 1;
     repeated TestInstancesGroupConfigError errors = 2;
-}
-
-// StatesService
-
-enum NoCloudState {
-    INIT = 0;
-    UNKNOWN = 1;
-    STOPPED = 2;
-    RUNNING = 3;
-    FAILURE  = 4;
-    DELETED = 5;
-    SUSPENDED = 6;
-    OPERATION = 7;
-}
-
-message State {
-    NoCloudState state = 1; // NoCloud Instance State
-    map<string, google.protobuf.Value> meta = 2; // Driver(Provider/Hypervisor) State data
-}
-
-message PostStateRequest {
-    string uuid = 1;
-    State state = 2;
-}
-message PostStateResponse {
-    string uuid = 1;
-    int64 result = 2;
-    string error = 3;
-}
-
-message GetStatesRequest {
-    repeated string instances = 1;
-}
-message GetStatesResponse {
-    map<string, State> states = 1; // Map of Instance UUID to it's State
-}
-
-service StatesService {
-    rpc PostState(PostStateRequest) returns (PostStateResponse) {};
-    rpc GetStates(GetStatesRequest) returns (GetStatesResponse) {};
 }
