+++ pkg/services_providers/proto/services_providers.proto
@@ -18,6 +18,7 @@ package nocloud.services_providers;
 
 import "google/protobuf/struct.proto";
 import "google/api/annotations.proto";
+import "pkg/states/proto/states.proto";
 
 message Var {
     map<string, google.protobuf.Value> value = 1;
@@ -30,6 +31,7 @@ message ServicesProvider {
     map<string, google.protobuf.Value> secrets = 4; // json
     map<string, Var> vars = 5;
     map<string, google.protobuf.Struct> extentions = 6;
+    nocloud.states.State state = 7;
 }
 
 message TestResponse {
